Resolve weird issues looking up addresses by switching interfaceIndex to the proper type

+++ Bonjour/MYAddressLookup.m
@@ -24,7 +24,7 @@
 {
     MYBonjourService *_service;
     NSString *_hostname;
-    UInt16 _interfaceIndex;
+    uint32_t _interfaceIndex;
     NSMutableSet *_addresses;
     UInt16 _port;
     CFAbsoluteTime _expires;
